Structure

Radiant.Elastic_Collision — Type

Elastic_Collision

Structure used to define parameters for production of multigroup elastic collision cross-sections.

Mandatory field(s)

N/A

Optional field(s) - with default values

interaction_types::Dict{Tuple{DataType,DataType},Vector{String}} = Dict((Electron,Electron) => ["S"],(Positron,Positron) => ["S"]) : Dictionary of the interaction processes types, of the form (incident particle,outgoing particle) => associated list of interaction type, which values correspond:

(Electron,Electron) => ["S"] : elastic interaction of electrons.

(Positron,Positron) => ["S"] : elastic interaction of positrons.

Methods

Radiant.set_model — Method

set_model(this::Elastic_Collision,model::String,is_KC::Bool=true)

To define the elastic scattering model.

Input Argument(s)

this::Elastic_Collision : elastic collisions structure.

model::String : model of elastic scattering:

rutherford : screened Rutherford cross-sections.

boschini : screened Mott cross-sections.

is_KC::Bool : Apply Karakow correction (true) or not (false).

is_SC::Bool : Apply Seltzer correction (true) or not (false).

Output Argument(s)

N/A

Examples

julia> elastic_collisions = Elastic_Collision()
julia> elastic_collisions.set_model("rutherford")

Radiant.set_transport_correction — Method

set_transport_correction(this::Elastic_Collision,is_ETC::Bool)

Enable or not extended transport correcton.

Input Argument(s)

this::Elastic_Collision : elastic collisions structure.

is_ETC::Bool : Enable (true) or not (false) extended transport correcton.

Output Argument(s)

N/A

Examples

julia> elastic_collisions = Elastic_Collision()
julia> elastic_collisions.is_ETC(false)

Radiant.set_solver — Method

set_solver(this::Elastic_Collision,solver::String)

Dictate how the cross-sections is distributed to the Boltzmann and the Fokker-Planck operators.

Input Argument(s)

this::Elastic_Collision : elastic collisions structure.

solver::String : model of elastic scattering:

BTE : the cross-sections are made to be used with the Boltzmann operator.

FP : the cross-sections are made to be used with the Fokker-Planck operator.

BFP : the cross-sections are distributed to the Boltzmann and Fokker-Planck operator.

Output Argument(s)

N/A

Examples

julia> elastic_collisions = Elastic_Collision()
julia> elastic_collisions.is_AFP(false)

Radiant.set_interaction_types — Method

set_interaction_types(this::Elastic_Collision,interaction_types::Dict{Tuple{DataType,DataType},Vector{String}})

To define the interaction types for Elastic_Collision processes.

Input Argument(s)

this::Elastic_Collision : elastic collisions structure.

interaction_types::Dict{Tuple{DataType,DataType},Vector{String}} : Dictionary of the interaction processes types, of the form (incident particle,outgoing particle) => associated list of interaction type, which can be:

(Electron,Electron) => ["S"] : elastic interaction of electrons.

(Positron,Positron) => ["S"] : elastic interaction of positrons.

Output Argument(s)

N/A

Examples

julia> elastic_collisions = Elastic_Collision()
julia> elastic_collisions.set_interaction_types( Dict((Positron,Positron) => ["S"]) ) # Elastic only for positrons