Fixed_Sources

Structure

Radiant.Fixed_Sources — Type

Fixed_Sources

Structure used to define a collection of fixed sources and their properties.

Mandatory field(s)

sources_list::Vector{Source} : list of VolumeSource or SurfaceSource sources.

cross_sections : Cross_Sections structure.

geometry : Geometry structure.

solvers : Solvers structure.

Optional field(s) - with default values

N/A

Build

Radiant.build — Method

build(this::Fixed_Sources)

To build the fixed sources structure.

Input Argument(s)

this::Fixed_Sources : fixed sources.

Output Argument(s)

N/A

Examples

julia> fs = Fixed_Sources()
julia> ... # Defining the fixed sources properties
julia> fs.build()

Setters

Radiant.add_source — Method

add_source(this::Fixed_Sources,fixed_source::Union{Surface_Source,Volume_Source})

To add a surface or volume source to the Fixed_Sources structure.

Input Argument(s)

this::Fixed_Sources : collection of fixed sources.

fixed_source::Union{Surface_Source,Volume_Source} : volume or surface source.

Output Argument(s)

N/A

Examples

julia> vs = Volume_Source()
julia> ... # Define the volume source properties
julia> fs = Fixed_Sources()
julia> fs.add_source(vs)

Getters

Radiant.get_source — Method

get_source(this::Fixed_Sources,particle::Particle)

Get the sources for a given particle.

Input Argument(s)

this::Fixed_Sources : collection of fixed sources.

particle::Particle : particle.

Output Argument(s)

source::Source : sources for the given particle.

Radiant.get_particles — Method

get_particles(this::Fixed_Sources)

Get the particle list.

Input Argument(s)

this::Fixed_Sources : collection of fixed sources.

Output Argument(s)

particles::Vector{Particle} : particle list.

Radiant.get_normalization_factor — Method

get_normalization_factor(this::Fixed_Sources)

Get the normalization factor.

Input Argument(s)

this::Fixed_Sources : collection of fixed sources.

Output Argument(s)

normalization_factor::Float64 : normalization factor.