Structure

Radiant.Inelastic_Leptons — Type

Inelastic_Leptons

Structure used to define parameters for production of multigroup inelastic collisional cross-sections for leptons.

Mandatory field(s)

N/A

Optional field(s) - with default values

interaction_types::Dict{Tuple{DataType,DataType},Vector{String}} = Dict((Positron,Positron) => ["S"],(Positron,Electron) => ["P"],(Electron,Electron) => ["S","P"]) : Dictionary of the interaction processes types, of the form (incident particle,outgoing particle) => associated list of interaction type, which values correspond:

(Positron,Positron) => ["S"] : scattering of incident positrons

(Electron,Electron) => ["S"] : scattering of incident electrons

(Positron,Electron) => ["P"] : production of electrons by incident positrons

(Electron,Electron) => ["P"] : production of electrons by incident electrons

Methods

Radiant.set_density_correction — Method

set_density_correction(this::Inelastic_Leptons,density_correction::String)

Set the Fermi density correction.

Input Argument(s)

this::Inelastic_Leptons : inelastic leptons structure.

density_correction::String : type of density effect:

fano : Fano density effect.

sternheimer : Sternheimer semi-empirical density effect.

none : no density effect.

Output Argument(s)

N/A

Examples

julia> elastic_leptons = Inelastic_Leptons()
julia> elastic_leptons.set_density_correction("sternheimer")

Radiant.set_is_shell_correction — Method

set_is_shell_correction(this::Inelastic_Leptons,is_shell_correction::Bool)

Activate or desactivate shell correction for stopping powers.

Input Argument(s)

this::Inelastic_Leptons : inelastic leptons structure.

is_shell_correction::Bool : activate (true) or desactivate (false) the shell correction.

Output Argument(s)

N/A

Examples

julia> elastic_leptons = Inelastic_Leptons()
julia> elastic_leptons.set_is_shell_correction(false)

Radiant.set_is_subshells_dependant — Method

set_is_subshells_dependant(this::Inelastic_Leptons,is_subshells_dependant::Bool)

Compute the inelastic cross-sections assuming bounded or unbounded electrons.

Input Argument(s)

this::Inelastic_Leptons : inelastic leptons structure.

is_subshells_dependant::Bool : bounded (true) or unbounded (false) electrons.

Output Argument(s)

N/A

Examples

julia> elastic_leptons = Inelastic_Leptons()
julia> elastic_leptons.set_is_subshells_dependant(false)

Radiant.set_interaction_types — Method

set_interaction_types(this::Inelastic_Leptons,interaction_types::Dict{Tuple{DataType,DataType},Vector{String}})

To define the interaction types for inelastic leptons processes.

Input Argument(s)

this::Inelastic_Leptons : inelastic leptons structure.

interaction_types::Dict{Tuple{DataType,DataType},Vector{String}} : Dictionary of the interaction processes types, of the form (incident particle,outgoing particle) => associated list of interaction type, which can be:

(Positron,Positron) => ["S"] : scattering of incident positrons

(Electron,Electron) => ["S"] : scattering of incident electrons

(Positron,Electron) => ["P"] : production of electrons by incident positrons

(Electron,Electron) => ["P"] : production of electrons by incident electrons

Output Argument(s)

N/A

Examples

julia> elastic_leptons = Inelastic_Leptons()
julia> elastic_leptons.set_interaction_types( Dict((Electron,Electron) => ["S"]) ) # No knock-on electrons.