Structure

Radiant.Photoelectric — Type

Photoelectric

Structure used to define parameters for production of multigroup photoelectric cross-sections.

Mandatory field(s)

N/A

Optional field(s) - with default values

interaction_types::Dict{Tuple{DataType,DataType},Vector{String}} = Dict((Photon,Photon) => ["A"],(Photon,Electron) => ["P"]) : Dictionary of the interaction processes types, of the form (incident particle,outgoing particle) => associated list of interaction type, which values correspond:

(Photon,Photon) => ["A"] : absorption of incoming photon.

(Photon,Electron) => ["P"] : produced photo-electron.

Methods

Radiant.set_model — Method

set_model(this::Photoelectric,model::String)

To define the photoelectric model.

Input Argument(s)

this::Photoelectric : photoelectric structure.

model::String : cross-section model:

jendl5 : evaluated subshell-dependent cross-sections.

biggs_lighthill : Biggs and Lighthill cross-sections.

Output Argument(s)

N/A

Examples

julia> photoelectric = Photoelectric()
julia> photoelectric.set_model("biggs_lighthill")

Radiant.set_interaction_types — Method

set_interaction_types(this::Photoelectric,interaction_types::Dict{Tuple{DataType,DataType},Vector{String}})

To define the interaction types for photoelectric processes.

Input Argument(s)

this::Photoelectric : photoelectric structure.

interaction_types::Dict{Tuple{DataType,DataType},Vector{String}} : Dictionary of the interaction processes types, of the form (incident particle,outgoing particle) => associated list of interaction type, which can be:

(Photon,Photon) => ["A"] : absorption of incoming photon.

(Photon,Electron) => ["P"] : produced photo-electron.

Output Argument(s)

N/A

Examples

julia> photoelectric = Photoelectric()
julia> photoelectric.set_interaction_types( Dict((Photon,Photon) => ["A"],(Photon,Electron) => ["P"]) ) # Full photoelectric phenomenon (default case).