Solvers

A Solvers object groups one discretization method per particle and controls the coupling between them.

Structure

Radiant.Solvers — Type

Solvers

Structure used to define the collection of discretization methods for transport calculations associated with each of the particle and additionnal coupled transport informations.

Mandatory field(s)

methods_list::Vector{SN} : list of the particle methods

maximum_number_of_generations::Int64 : number of particle generations to transport.

Optional field(s) - with default values

N/A

Setters

Radiant.add_solver — Method

add_solver(this::Solvers,method::Solver)

To add a particle and is associated methods to the Solvers structure.

Input Argument(s)

this::Solvers : collection of discretization method.

method::Solver : discretization method.

Output Argument(s)

N/A

Examples

julia> m = SN()
julia> ... # Define the methods properties
julia> ms = Solvers()
julia> ms.add_solver(m)

Radiant.set_maximum_number_of_generations — Method

set_maximum_number_of_generations(this::Solvers,maximum_number_of_generations::Int64)

To set the maximum number of particle generation to transport during calculations.

Input Argument(s)

this::Solvers : collection of discretization method.

maximum_number_of_generations::Int64 : maximum number of particle generation to transport during calculations.

Output Argument(s)

N/A

Examples

julia> ms = Solvers()
julia> ms.set_maximum_number_of_generations(2)

Radiant.set_convergence_criterion — Method

set_convergence_criterion(this::Solvers,convergence_criterion::Real)

To set the convergence criterion for coupled particle transport.

Input Argument(s)

this::Solvers : collection of discretization method.

convergence_criterion::Real : convergence criterion.

Output Argument(s)

N/A

Examples

julia> ms = Solvers()
julia> ms.set_convergence_criterion(1e-7)

Radiant.set_convergence_type — Method

set_convergence_type(this::Solvers,convergence_type::String)

To set the convergence type.

Input Argument(s)

this::Solvers : collection of discretization method.

convergence_type::String : convergence type, which can be:

convergence_type = "flux" : convergence of the fluxes per particle.

convergence_type = "energy-deposition" : convergence of the total energy deposition.

convergence_type = "charge-deposition" : convergence of the total charge deposition.

Output Argument(s)

N/A

Examples

julia> ms = Solvers()
julia> ms.set_convergence_type("energy-deposition")

Getters

Radiant.get_method — Method

get_method(this::Solvers,particle::Particle)

Get the method associated with a particle.

Input Argument(s)

this::Solvers : collection of discretization method.

particle::Particle : particle.

Output Argument(s)

method::Solver : method associated with the particle.

Radiant.get_particles — Method

get_particles(this::Solvers)

Get the particle list.

Input Argument(s)

this::Solvers : collection of discretization method.

Output Argument(s)

particle::Vector{Particle} : particle list.

Radiant.get_number_of_particles — Method

get_number_of_particles(this::Solvers)

Get the number of particles.

Input Argument(s)

this::Solvers : collection of discretization method.

Output Argument(s)

number_of_particles::Int64 : number of particles.

Radiant.get_maximum_number_of_generations — Method

get_maximum_number_of_generations(this::Solvers)

Get the maximum number of generations.

Input Argument(s)

this::Solvers : collection of discretization method.

Output Argument(s)

maximum_number_of_generations::Int64 : maximum number of generations.

Radiant.get_convergence_criterion — Method

get_convergence_criterion(this::Solvers)

Get the convergence criterion.

Input Argument(s)

this::Solvers : collection of discretization method.

Output Argument(s)

convergence_criterion::Float64 : convergence criterion.

Radiant.get_convergence_type — Method

get_convergence_type(this::Solvers)

Get the convergence type.

Input Argument(s)

this::Solvers : collection of discretization method.

Output Argument(s)

convergence_type::String : convergence type.