Structure

Radiant.Surface_Source — Type

Surface_Source

Structure used to define a directionnal boundary source and its properties.

Mandatory field(s)

name::String : name (or identifier) of the Surface_Source structure.

particle::Particle : type of particle emitted.

energy_group::Int64 : energy group index in which the particle are emitted.

direction::Vector{Float64} : direction cosine.

location::String : boundary at which the source is located.

boundaries::Vector{Float64} : boundaries of the source along each axis [in cm].

Optional field(s) - with default values

intensity::Float64=1.0 : intensity [# particles/cm⁽ᴺ⁻¹⁾, where N is the geometry dimension].

Methods

Radiant.set_particle — Method

set_particle(this::Surface_Source,particle::Particle)

To define the source particle.

Input Argument(s)

this::Surface_Source : surface source.

particle::String : particle.

Output Argument(s)

N/A

Examples

julia> ss = Surface_Source()
julia> ss.set_particle(electron)

Radiant.set_intensity — Method

set_intensity(this::Surface_Source,intensity::Real)

To define the intensity of the source.

Input Argument(s)

this::Surface_Source : surface source.

intensity::Float64 : intensity [# particles/cmᴺ, where N is the geometry dimension]

Output Argument(s)

N/A

Examples

julia> ss = Surface_Source()
julia> ss.set_intensity(100)

Radiant.set_energy_group — Method

set_energy_group(this::Surface_Source,energy_group::Int64)

To define the energy of the source by setting the energy group in which they are produced.

Input Argument(s)

this::Surface_Source : surface source.

energy_group::Int64 : energy group index.

Output Argument(s)

N/A

Examples

julia> ss = Surface_Source()
julia> ss.set_energy_group(1)

Radiant.set_direction — Method

set_direction(this::Surface_Source,direction::Vector{Float64})

To set a direction of the source.

Input Argument(s)

this::Surface_Source : surface source.

direction::Vector{Float64} : director cosines [μ,η,ξ]

Output Argument(s)

N/A

Examples

julia> ss = Surface_Source()
julia> ss.set_direction([1.0,0.0,0.0])

Radiant.set_location — Method

set_location(this::Surface_Source,location::String)

To set the location of the surface source.

Input Argument(s)

this::Surface_Source : surface source.

location::String : boundary on which the surface source is, which can takes the following value:

boundary = "x-" : the lower bound along x-axis

boundary = "x+" : the upper bound along x-axis

boundary = "y-" : the lower bound along y-axis

boundary = "y+" : the upper bound along y-axis

boundary = "z-" : the lower bound along z-axis

boundary = "z+" : the upper bound along z-axis

Output Argument(s)

N/A

Examples

julia> ss = Surface_Source()
julia> ss.set_location("x-")

Radiant.set_boundaries — Method

set_boundaries(this::Surface_Source,axis::String,boundaries::Vector{Float64})

To define the boundaries of the source along the specified axis.

Input Argument(s)

this::Surface_Source : surface source.

axis::String : axis along which the boundaries are defined.

boundaries::Vector{Float64} : boundaries of the source in accending order [in cm]

Output Argument(s)

N/A

Examples

julia> ss = Surface_Source()
julia> ss.set_boundaries("x",[1.0,3.0])