Structure

Radiant.Compton — Type

Compton

Structure used to define parameters for production of multigroup Compton cross-sections.

Mandatory field(s)

N/A

Optional field(s) - with default values

interaction_types::Dict{Tuple{String,String},Vector{String}} = Dict(("photons","photons") => ["S"],("photons","electrons") => ["P"]): Dictionary of the interaction processes types, of the form (incident particle,outgoing particle) => associated list of interaction type, which values correspond:

("photons","photons") => ["S"]: scattering of incident photon following Compton interaction.

("photons","electrons") => ["P"]: produced electron following Compton interaction.

Methods

Missing docstring.

Missing docstring for Radiant.set_is_waller_hartree_factor(this::Radiant.Compton,is_waller_hartree_factor::Bool). Check Documenter's build log for details.

Radiant.set_interaction_types — Method

set_interaction_types(this::Compton,interaction_types::Dict{Tuple{String,String},Vector{String}})

To define the interaction types for Compton processes.

Input Argument(s)

this::Compton: compton structure.

interaction_types::Dict{Tuple{String,String},Vector{String}}: Dictionary of the interaction processes types, of the form (incident particle,outgoing particle) => associated list of interaction type, which can be:

("photons","photons") => ["S"]: scattering of incident photon following Compton interaction.

("photons","electrons") => ["P"]: produced electron following Compton interaction.

Output Argument(s)

N/A

Examples

julia> compton = Compton()
julia> compton.set_interaction_types( Dict(("photons","photons") => ["S"]) ) # Electron are absorbed following Compton interaction.