Structure

Radiant.Computation_Unit — Type

Computation_Unit

Structure used to consolidate the cross-sections, geometry, methods and sources, execute transport calculations and extract its results.

User-defined field(s)

Mandatory field(s)

cross_sections::Cross_Sections: cross-section library.

geometry::Geometry: geometry.

methods::Methods: methods.

sources::Sources: fixed sources.

Optional field(s) - with default values

System-defined field(s)

- `flux`: flux solution.

Methods

Radiant.set_cross_sections — Method

set_cross_sections(this::Computation_Unit,cross_sections::Cross_Sections)

To set the cross-sections library for transport calculations.

Input Argument(s)

this::Computation_Unit: computation unit.

cross_sections::Cross_Sections: cross-sections library.

Output Argument(s)

N/A

Examples

julia> cs = Cross_Sections()
julia> ... # Define cross-sections properties and generate multigroup cross-sections.
julia> cu = Computation_Unit()
julia> cu.set_cross_sections(cs)

Radiant.set_geometry — Method

set_geometry(this::Computation_Unit,geometry::Geometry)

To set the geometry for transport calculations.

Input Argument(s)

this::Computation_Unit: computation unit.

geometry::Geometry: geometry.

Output Argument(s)

N/A

Examples

julia> geo = Geometry()
julia> ... # Define geometry and its properties
julia> cu = Computation_Unit()
julia> cu.set_geometry(geo)

Radiant.set_methods — Method

set_methods(this::Computation_Unit,methods::Methods)

To set the discretization methods for transport calculations.

Input Argument(s)

this::Computation_Unit: computation unit.

methods::Methods: collection of discretization method per particle.

Output Argument(s)

N/A

Examples

julia> ms = Methods()
julia> ... # Define all the discretization methods and their properties
julia> cu = Computation_Unit()
julia> cu.set_methods(ms)

Radiant.set_sources — Method

set_sources(this::Computation_Unit,sources::Fixed_Sources)

To set the fixed sources for transport calculations.

Input Argument(s)

this::Computation_Unit: computation unit.

sources::Fixed_Sources: collection of fixed sources.

Output Argument(s)

N/A

Examples

julia> fs = Fixed_Sources()
julia> ... # Define all the fixed sources and their properties
julia> cu = Computation_Unit()
julia> cu.set_sources(fs)

Radiant.run — Method

run(this::Computation_Unit)

To lauch the transport calculations to solve the transport equation and obtain the flux solution.

Input Argument(s)

this::Computation_Unit: computation unit.

Output Argument(s)

N/A

Examples

julia> cu = Computation_Unit()
julia> ... # Define the cross-sections, geometry, fixed sources and discretization methods
julia> cu.run()