Structure

Radiant.Elastic_Leptons — Type

Elastic_Leptons

Structure used to define parameters for production of multigroup elastic cross-sections for leptons.

Mandatory field(s)

N/A

Optional field(s) - with default values

interaction_types::Dict{Tuple{String,String},Vector{String}} = Dict(("electrons","electrons") => ["S"],("positrons","positrons") => ["S"]): Dictionary of the interaction processes types, of the form (incident particle,outgoing particle) => associated list of interaction type, which values correspond:

("electrons","electrons") => ["S"]: elastic interaction of electrons.

("positrons","positrons") => ["S"]: elastic interaction of positrons.

Methods

Radiant.set_model — Method

set_model(this::Elastic_Leptons,model::String,is_KC::Bool=true)

To define the elastic scattering model.

Input Argument(s)

this::Elastic_Leptons: elastic leptons structure.

model::String: model of elastic scattering:

rutherford: screened Rutherford cross-sections.

mott: screened Mott cross-sections.

is_KC::Bool: Apply Karakow's Correction (true) or not (false).

is_SC::Bool: Apply Seltzer's Correction (true) or not (false).

Output Argument(s)

N/A

Examples

julia> elastic_leptons = Elastic_Leptons()
julia> elastic_leptons.set_model("rutherford")

Radiant.set_transport_correction — Method

set_transport_correction(this::Elastic_Leptons,is_ETC::Bool)

Enable or not extended transport correcton.

Input Argument(s)

this::Elastic_Leptons: elastic leptons structure.

is_ETC::Bool: Enable (true) or not (false) extended transport correcton.

Output Argument(s)

N/A

Examples

julia> elastic_leptons = Elastic_Leptons()
julia> elastic_leptons.is_ETC(false)

Missing docstring.

Missing docstring for Radiant.set_angular_fokker_planck(this::Radiant.Elastic_Leptons,is_AFP::Bool). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Radiant.set_kawrakow_correction(this::Radiant.Elastic_Leptons,is_kawrakow_correction::Bool,subshell_dependant_inelastic::Bool=true). Check Documenter's build log for details.

Radiant.set_interaction_types — Method

set_interaction_types(this::Elastic_Leptons,interaction_types::Dict{Tuple{String,String},Vector{String}})

To define the interaction types for Elastic_Leptons processes.

Input Argument(s)

this::Elastic_Leptons: elastic leptons structure.

interaction_types::Dict{Tuple{String,String},Vector{String}}: Dictionary of the interaction processes types, of the form (incident particle,outgoing particle) => associated list of interaction type, which can be:

("electrons","electrons") => ["S"]: elastic interaction of electrons.

("positrons","positrons") => ["S"]: elastic interaction of positrons.

Output Argument(s)

N/A

Examples

julia> elastic_leptons = Elastic_Leptons()
julia> elastic_leptons.set_interaction_types( Dict(("positrons","positrons") => ["S"]) ) # Elastic only for positrons