Structure

Radiant.Solvers — Type

Solvers

Structure used to define the collection of discretization methods for transport calculations associated with each of the particle and additionnal coupled transport informations.

Mandatory field(s)

methods_list::Vector{Discrete_Ordinates}: list of the particle methods

number_of_generations::Int64: number of particle generations to transport.

Optional field(s) - with default values

N/A

Methods

Radiant.add_solver — Method

add_solver(this::Solvers,method::Discrete_Ordinates)

To add a particle and is associated methods to the Solvers structure.

Input Argument(s)

this::Solvers: collection of discretization method.

method::Discrete_Ordinates: discretization method.

Output Argument(s)

N/A

Examples

julia> m = Discrete_Ordinates()
julia> ... # Define the methods properties
julia> ms = Solvers()
julia> ms.add_solver(m)

Radiant.set_number_of_generations — Method

set_number_of_generations(this::Solvers,number_of_generations::Int64)

To set the number of particle generation to transport during calculations

Input Argument(s)

this::Solvers: collection of discretization method.

number_of_generations::Int64: number of particle generation to transport during calculations.

Output Argument(s)

N/A

Examples

julia> ms = Solvers()
julia> ms.set_number_of_generations(2)