Structure

Radiant.Volume_Source — Type

Volume_Source

Structure used to define an isotropic volume source and its properties.

User-defined field(s)

Mandatory field(s)

name::String: name (or identifier) of the Volume_Source structure.

particle::String: type of particle emitted.

energy_group::Int64: energy group index in which the particle are emitted.

boundaries::Vector{Float64}: boundaries of the source along each axis [in cm].

Optional field(s) - with default values

intensity::Float64=1.0: intensity [# particles/cmᴺ, where N is the geometry dimension].

System-defined field(s)

is_build::Bool: boolean value defining if the MultigroupCrossSections was build or not.

volume_sources::Array{Float64}: formatted array for transport calculations.

normalization_factor::Float64: normalization over the number of particles.

Methods

Radiant.set_particle — Method

set_particle(this::Volume_Source,particle::String)

To define the source particle.

Input Argument(s)

this::Volume_Source: volume source.

particle::String: type of particle, which can takes the following values:

particle = "photons": photons.

particle = "electrons": electrons.

particle = "positrons": positrons.

Output Argument(s)

N/A

Examples

julia> vs = Volume_Source()
julia> vs.set_particle("electrons")

Radiant.set_intensity — Method

set_intensity(this::Volume_Source,intensity::Real)

To define the intensity of the source.

Input Argument(s)

this::Volume_Source: volume source.

intensity::Float64: intensity [# particles/cmᴺ, where N is the geometry dimension]

Output Argument(s)

N/A

Examples

julia> vs = Volume_Source()
julia> vs.set_intensity(100)

Radiant.set_energy_group — Method

set_energy_group(this::Volume_Source,energy_group::Int64)

To define the energy of the source by setting the energy group in which they are produced.

Input Argument(s)

this::Volume_Source: volume source.

energy_group::Int64: energy group index.

Output Argument(s)

N/A

Examples

julia> vs = Volume_Source()
julia> vs.set_energy_group(1)

Radiant.set_boundaries — Method

set_boundaries(this::Volume_Source,axis::String,boundaries::Vector{Float64})

To define the boundaries of the source along the specified axis.

Input Argument(s)

this::Volume_Source: volume source.

axis::String: axis along which the boundaries are defined.

boundaries::Vector{Float64}: boundaries of the source in accending order [in cm]

Output Argument(s)

N/A

Examples

julia> vs = Volume_Source()
julia> vs.set_boundaries("x",[1.0,3.0])