Structure

Radiant.Fixed_Sources — Type

Fixed_Sources

Structure used to define a collection of fixed sources and their properties.

Mandatory field(s)

sources_list::Vector{Source} : list of VolumeSource or SurfaceSource sources.

cross_sections : Cross_Sections structure.

geometry : Geometry structure.

solvers : Solvers structure.

Optional field(s) - with default values

N/A

Methods

Radiant.add_source — Method

add_source(this::Fixed_Sources,fixed_source::Union{Surface_Source,Volume_Source})

To add a surface or volume source to the Fixed_Sources structure.

Input Argument(s)

this::Fixed_Sources : collection of fixed sources.

fixed_source::Union{Surface_Source,Volume_Source} : volume or surface source.

Output Argument(s)

N/A

Examples

julia> vs = Volume_Source()
julia> ... # Define the volume source properties
julia> fs = Fixed_Sources()
julia> fs.add_source(vs)