Structure

Radiant.Solvers — Type

Solvers

Structure used to define the collection of discretization methods for transport calculations associated with each of the particle and additionnal coupled transport informations.

Mandatory field(s)

methods_list::Vector{Discrete_Ordinates} : list of the particle methods

maximum_number_of_generations::Int64 : number of particle generations to transport.

Optional field(s) - with default values

N/A

Methods

Radiant.add_solver — Method

add_solver(this::Solvers,method::Solver)

To add a particle and is associated methods to the Solvers structure.

Input Argument(s)

this::Solvers : collection of discretization method.

method::Solver : discretization method.

Output Argument(s)

N/A

Examples

julia> m = Discrete_Ordinates()
julia> ... # Define the methods properties
julia> ms = Solvers()
julia> ms.add_solver(m)

Radiant.set_maximum_number_of_generations — Method

set_maximum_number_of_generations(this::Solvers,maximum_number_of_generations::Int64)

To set the maximum number of particle generation to transport during calculations.

Input Argument(s)

this::Solvers : collection of discretization method.

maximum_number_of_generations::Int64 : maximum number of particle generation to transport during calculations.

Output Argument(s)

N/A

Examples

julia> ms = Solvers()
julia> ms.set_maximum_number_of_generations(2)

Radiant.set_convergence_criterion — Method

set_convergence_criterion(this::Solvers,convergence_criterion::Real)

To set the convergence criterion for coupled particle transport.

Input Argument(s)

this::Solvers : collection of discretization method.

convergence_criterion::Real : convergence criterion.

Output Argument(s)

N/A

Examples

julia> ms = Solvers()
julia> ms.set_convergence_criterion(1e-7)

Radiant.set_convergence_type — Method

set_convergence_type(this::Solvers,convergence_type::String)

To set the convergence type.

Input Argument(s)

this::Solvers : collection of discretization method.

convergence_type::String : convergence type, which can be:

convergence_type = "flux" : convergence of the fluxes per particle.

convergence_type = "energy-deposition" : convergence of the total energy deposition.

convergence_type = "charge-deposition" : convergence of the total charge deposition.

Output Argument(s)

N/A

Examples

julia> ms = Solvers()
julia> ms.set_convergence_type("energy-deposition")