Structure

Radiant.Computation_Unit — Type

Computation_Unit

Structure used to consolidate the cross-sections, geometry, solvers and sources, execute transport calculations and extract its results.

Mandatory field(s)

cross_sections::Cross_Sections: cross-section library.

geometry::Geometry: geometry.

solvers::Solvers: solvers.

sources::Sources: fixed sources.

Optional field(s) - with default values

N/A

Methods

Radiant.set_cross_sections — Method

set_cross_sections(this::Computation_Unit,cross_sections::Cross_Sections)

To set the cross-sections library for transport calculations.

Input Argument(s)

this::Computation_Unit: computation unit.

cross_sections::Cross_Sections: cross-sections library.

Output Argument(s)

N/A

Examples

julia> cs = Cross_Sections()
julia> ... # Define cross-sections properties and generate multigroup cross-sections.
julia> cu = Computation_Unit()
julia> cu.set_cross_sections(cs)

Radiant.set_geometry — Method

set_geometry(this::Computation_Unit,geometry::Geometry)

To set the geometry for transport calculations.

Input Argument(s)

this::Computation_Unit: computation unit.

geometry::Geometry: geometry.

Output Argument(s)

N/A

Examples

julia> geo = Geometry()
julia> ... # Define geometry and its properties
julia> cu = Computation_Unit()
julia> cu.set_geometry(geo)

Radiant.set_solvers — Method

setsolvers(this::ComputationUnit,solvers::Solvers)

To set the solvers for transport calculations.

Input Argument(s)

this::Computation_Unit: computation unit.

solvers::Solvers: collection of solvers per particle.

Output Argument(s)

N/A

Examples

julia> ms = Solvers()
julia> ... # Define all the discretization solvers and their properties
julia> cu = Computation_Unit()
julia> cu.set_solvers(ms)

Radiant.set_sources — Method

set_sources(this::Computation_Unit,sources::Fixed_Sources)

To set the fixed sources for transport calculations.

Input Argument(s)

this::Computation_Unit: computation unit.

sources::Fixed_Sources: collection of fixed sources.

Output Argument(s)

N/A

Examples

julia> fs = Fixed_Sources()
julia> ... # Define all the fixed sources and their properties
julia> cu = Computation_Unit()
julia> cu.set_sources(fs)

Radiant.run — Method

run(this::Computation_Unit)

To lauch the transport calculations to solve the transport equation and obtain the flux solution.

Input Argument(s)

this::Computation_Unit: computation unit.

Output Argument(s)

N/A

Examples

julia> cu = Computation_Unit()
julia> ... # Define the cross-sections, geometry, fixed sources and discretization solvers
julia> cu.run()

Radiant.get_voxels_position — Method

get_voxels_position(this::Computation_Unit,axis::String)

To set the mid-point voxels position along the specified axis.

Input Argument(s)

this::Computation_Unit: computation unit.

axis::String: axis, which can takes the following values:

boundary = "x": along x-axis

boundary = "y": along y-axis

boundary = "z": along z-axis

Output Argument(s)

x::Vector{Float64}: mid-point voxels position along the specified axis.

Examples

julia> cu = Computation_Unit()
julia> ... # Define computation unit and run it.
julia> x = cu.get_voxels_position("x")

Radiant.get_energies — Method

get_energies(this::Computation_Unit,particle::String)

To set the mid-point energy in each group for the specified particle.

Input Argument(s)

this::Computation_Unit: computation unit.

particle::String: particle identifier, where each particle is either:

particle = "photons": photons.

particle = "electrons": electrons.

particle = "positrons": positrons.

Output Argument(s)

E::Vector{Float64}: mid-point energy in each group for the specified particle.

Examples

julia> cu = Computation_Unit()
julia> ... # Define computation unit and run it.
julia> E = cu.get_voxels_position("electrons")

Radiant.get_flux — Method

get_flux(this::Computation_Unit,particle::String)

To get the array containing the flux in each voxels and in each energy group for the specified particle.

Input Argument(s)

this::Computation_Unit: computation unit.

particle::String: flux particle, which can takes the following values:

particle = "photons": photons charge deposition.

particle = "electrons": electrons charge deposition.

particle = "positrons": positrons charge deposition.

Output Argument(s)

flux::Array{Float64}: flux array.

Examples

julia> cu = Computation_Unit()
julia> ... # Define computation unit and run it.
julia> flux = cu.get_flux("electrons")

Radiant.get_energy_deposition — Method

get_energy_deposition(this::Computation_Unit,type::String)

To get the array containing the energy deposition in each voxels.

Input Argument(s)

this::Computation_Unit: computation unit.

type::String: type of energy deposition, which can takes the following values:

type = "total": total energy deposition.

type = "photons": photons energy deposition.

type = "electrons": electrons energy deposition.

type = "positrons": positrons energy deposition.

Output Argument(s)

energy_deposition::Array{Float64}: energy deposition array.

Examples

julia> cu = Computation_Unit()
julia> ... # Define computation unit and run it.
julia> energy_deposition = cu.get_energy_deposition("total")

Radiant.get_charge_deposition — Method

get_charge_deposition(this::Computation_Unit,type::String)

To get the array containing the charge deposition in each voxels.

Input Argument(s)

this::Computation_Unit: computation unit.

type::String: type of charge deposition, which can takes the following values:

type = "total": total charge deposition.

type = "photons": photons charge deposition.

type = "electrons": electrons charge deposition.

type = "positrons": positrons charge deposition.

Output Argument(s)

charge_deposition::Array{Float64}: charge deposition array.

Examples

julia> cu = Computation_Unit()
julia> ... # Define computation unit and run it.
julia> charge_deposition = cu.get_charge_deposition("total")